app/requirements.txt
pyvista>=0.42.0
numpy>=1.20.0
gmsh>=4.11.0
pyyaml>=6.0
